requirements.txt
fastapi
uvicorn
pymongo
redis
kafka-python
scikit-learn
pandas
numpy
python-dotenv
